Merge to staging for 3.1.0 release

+++ kbase-extension/static/kbase/js/widgets/narrative_core/kbaseNarrativeWorkspace.js
@@ -1,5 +1,6 @@
-/*global define*/
+/*global define, document*/
 /*jslint white:true,browser:true,devel:true*/
+/*eslint-env browser*/
 
 /**
  * Top-level 'widget' for the workspace interaction with the KBase narrative.
@@ -43,8 +44,7 @@ define([
     'kbaseTabs',
     'common/props',
     'kb_service/client/narrativeMethodStore',
-
-    'bootstrap',
+    'bootstrap'
 ], function(
     Jupyter,
     Runtime,
@@ -146,7 +146,7 @@ define([
             );
 
             $(document).on('dataUpdated.Narrative',
-                function(event) {
+                function() {
                     if (Jupyter && Jupyter.notebook) {
                         // XXX: This is a hell of a hack. I hate
                         // using the 'first time' bit like this,
@@ -241,7 +241,7 @@ define([
             // Global functions for setting icons
             $(document).on('setDataIcon.Narrative',
                 function(e, param) {
-                    this.setDataIcon(param.elt, param.type);
+                    this.setDataIcon(param.elt, param.type, param.stacked, param.indent);
                 }.bind(this)
             );
 
@@ -472,48 +472,38 @@ define([
                     appSpec: spec
                 }
             });
+            return cell;
         },
 
+
+        /*
+        For now we need to keep the "spec grokking" in place.
+        A little bit like duck typing, we inspect the properties of the app
+        spec to determine if it is an app, editor, or viewer.
+        */
         determineMethodCellType: function(spec) {
+
+            // TODO: remove this switch when merging in grouped paramters and
+            // legacy parameters.
+            //if (spec.parameter_groups && spec.parameter_groups.length > 0) {
+            //   return 'app2';
+            //}
+
             // An app will execute via the method described in the behavior. If
             // such a method is not described, it is by definition not an
             // executing app.
-            if ((spec.behavior.kb_service_method && spec.behavior.kb_service_name) ||
-                (spec.behavior.script_module && spec.behavior.script_name)) {
-                return 'app';
+            if (spec.behavior.kb_service_name && spec.behavior.kb_service_method) {
+                switch (spec.info.app_type) {
+                    case 'app':
+                        return 'app';
+                    case 'editor':
+                        return 'editor';
+                    default:
+                        throw new Error('This app does not specify spec.info.app_type');
+                }
             }
-
-            // The category property is supposedly used to indicate that the app
-            // is a viewer, but this is not used very reliably.
-            // Still, we look at that here...
-            if (spec.info.categories.some(function(category) {
-                    return (category === 'viewers');
-                })) {
-                return 'view';
-            }
-
-            // A very small class of methods are just non-app-calling widgets.
-            switch (spec.info.id) {
-                //case 'model_support/edit_model':
-                case 'fba_tools/edit_metabolic_model':
-                case 'fba_tools/create_or_edit_media':
-                //case 'model_support/edit_media':
-                    return 'widget';
-            }
-
-            // ... while in reality, ANY app which does not execute is for now
-            // considered a viewer.
 
             return 'view';
-
-            //            if (!spec.parameters.some(function (parameter) {
-            //                return (parameter.ui_class === 'output');
-            //            })) {
-            //                return 'app';
-            //            };
-            //
-            //            console.error('ERROR - could not determine cell type', spec);
-            //            throw new Error('Could not determine cell type');
         },
 
 
@@ -1517,7 +1507,7 @@ define([
          * just look to see if it is indeed a kbase cell, and if so we punt
          * to it.
          */
-        deleteCell: function (index) {
+        deleteCell: function(index) {
             if (index === undefined || index === null) {
                 return;
             }
@@ -1530,31 +1520,31 @@ define([
             var p = html.tag('p');
 
             if (!kbaseCellType || !cellId) {
-                UI.make({node: this.$elem[0]}).showConfirmDialog({
-                    title: 'Confirm Cell Deletion',
-                    body: [
-                        p('Cell deletion is permanent. There is no "undo" feature to recover this cell once it is deleted.'),
-                        p('Are you sure you want to delete this cell?')
-                    ]
-                })
-                .then(function (confirmed) {
-                    if (confirmed) {
-                        if (kbaseCellType && !cellId) {
-                            console.warn('KBase cell without cell id, DELETING ANYWAY!', cell.metadata);
+                UI.make({ node: this.$elem[0] }).showConfirmDialog({
+                        title: 'Confirm Cell Deletion',
+                        body: [
+                            p('Cell deletion is permanent. There is no "undo" feature to recover this cell once it is deleted.'),
+                            p('Are you sure you want to delete this cell?')
+                        ]
+                    })
+                    .then(function(confirmed) {
+                        if (confirmed) {
+                            if (kbaseCellType && !cellId) {
+                                console.warn('KBase cell without cell id, DELETING ANYWAY!', cell.metadata);
+                            }
+                            Jupyter.notebook.delete_cell(index);
                         }
-                        Jupyter.notebook.delete_cell(index);
-                    }
-                });
+                    });
                 return;
             }
 
             this.runtime.bus().send({}, {
-               channel: {
-                   cell: cellId
-               },
-               key: {
-                   type: 'delete-cell'
-               }
+                channel: {
+                    cell: cellId
+                },
+                key: {
+                    type: 'delete-cell'
+                }
             });
         },
 
@@ -2524,34 +2514,80 @@ define([
          *
          * @param $logo - Target element
          * @param type - Name of data type
+         * @param stacked - If true, show "stacked" version of the icon
+         *                indicating, e.g., that this is a container for
+         *                multiple items. Undefined is false.
+         * @param indent - Indent level (default is none)
          */
-        setDataIcon: function($logo, type) {
-            if ($logo.hasClass('exampleDataIcon')) {
-                console.debug("SET EXAMPLE ICON");
+        setDataIcon: function($logo, type, stacked, indent) {
+            if (indent === undefined || indent === null) {
+                indent = 0;
             }
+
             var icons = this.data_icons;
-            var icon = _.has(icons, type) ? icons[type] : icons['DEFAULT'];
+            var icon = _.has(icons, type) ? icons[type] : icons.DEFAULT;
             // background circle
             $logo.addClass("fa-stack fa-2x").css({
-                    'cursor': 'pointer'
-                })
-                .append($('<i>')
-                    .addClass("fa fa-circle fa-stack-2x")
-                    .css({
-                        'color': this.logoColorLookup(type)
-                    }));
+                'cursor': 'pointer'
+            });
+            // For 'stacked' (set) icons, add a shifted-over
+            // circle first, as the bottom layer, then also add a border
+            // to the top one.
+            var circle_classes = 'fa fa-circle fa-stack-2x';
+            var circle_color = this.logoColorLookup(type);
+            var cmax = function(x) { return x > 255 ? 255 : x; };
+            if (stacked) {
+                var parsed_color, r, g, b;
+                var cstep = 20; // color-step for overlapped circles
+                var num_stacked_circles = 1; // up to 2
+                // XXX: Assume color is in form '#RRGGBB'
+                if (circle_color[0] == '#') {
+                    parsed_color = circle_color.match(/#(..)(..)(..)/);
+                    r = parseInt(parsed_color[1], 16);
+                    g = parseInt(parsed_color[2], 16);
+                    b = parseInt(parsed_color[3], 16);
+                }
+                // XXX: Assume color is in form "rgb(#,#,#)"
+                else {
+                    parsed_color = circle_color.match(/rgb\s*\(\s*(\d+)\s*,\s*(\d+)\s*,\s*(\d+)\s*\)/);
+                    r = parsed_color[1];
+                    g = parsed_color[2];;
+                    b = parsed_color[3];
+                }
+                // Add circles with lighter colors
+                for (var i = num_stacked_circles; i > 0; i--) {
+                    var stacked_color = 'rgb(' + cmax(r + i * cstep) + ',' +
+                        cmax(g + i * cstep) + ',' + cmax(b + i * cstep) + ')';
+                    $logo.append($('<i>')
+                        .addClass(circle_classes + ' kb-data-list-logo-shiftedx' + i)
+                        .css({ 'color': stacked_color }));
+                    $logo.append($('<i>')
+                        .addClass(circle_classes + ' kb-data-list-logo-shifted' + i)
+                        .css({ 'color': 'white' }));
+                }
+            }
+            // Assume there are CSS rules for levels of indent we care about..
+            if (indent > 0) {
+                $logo.addClass('kb-data-list-level1');
+            } else if ($logo.hasClass('kb-data-list-level1')) {
+                $logo.removeClass('kb-data-list-level1');
+            }
+
+            $logo.append($('<i>')
+                .addClass(circle_classes)
+                .css({ 'color': circle_color }));
+            // to avoid repetition, define the func. here that will
+            // add one set of icons
+            var add_logo_func = function(fa_icon, $logo, cls) {
+                $logo.append($('<i>')
+                    .addClass(fa_icon + ' fa-inverse fa-stack-1x ' + cls));
+            };
             if (this.isCustomIcon(icon)) {
                 // add custom icons (more than 1 will look weird, though)
-                _.each(icon, function(cls) {
-                    $logo.append($('<i>')
-                        .addClass("icon fa-inverse fa-stack-1x " + cls));
-                });
+                _.each(icon, function(cls) { add_logo_func('icon', $logo, cls); });
             } else {
                 // add stack of font-awesome icons
-                _.each(icon, function(cls) {
-                    $logo.append($('<i>')
-                        .addClass("fa fa-inverse fa-stack-1x " + cls));
-                });
+                _.each(icon, function(cls) { add_logo_func('fa', $logo, cls); });
             }
         },
 
@@ -2572,9 +2608,9 @@ define([
          * @param type
          * @returns {string} Color code
          */
-        logoColorLookup: function (type) {
+        logoColorLookup: function(type) {
             var color = this.icon_color_mapping[type];
-            if ( color === undefined) {
+            if (color === undefined) {
                 // fall back to primitive hack that just guesses
                 var code = 0;
                 for (var i = 0; i < type.length; code += type.charCodeAt(i++));
